- Fixed Class & Files names for test files - Cleaned up the temporary debug logs - Replacing compile task with resolve task since cacheing is not effective in overall build time - Increased travis build partitions to 5
