+++ .travis.yml
@@ -25,18 +25,18 @@ cache:
     - $BUILD_CACHE_DIR
 jobs:
   include:
-    - stage: 'clone repo and compile project'
-      name: 'Clone repo & compile project'
+    - stage: 'clone repo and resolve dependencies'
+      name: 'Clone repo & resolve dependencies'
       script: |
-        find **/src/test/scala/** -type f -name '*.scala' | cut -d "/" -f5-$2 | sed 's/.\{6\}$/\*/g' > $TESTS_FILE_NAME
+        find **/src/test/scala/* -type f -name '*.scala' | cut -d "/" -f5-$2 | sed 's/.\{6\}$/\*/g' > $TESTS_FILE_NAME
         TOTAL_TESTS=`cat $TESTS_FILE_NAME | wc -l`
         pushd $BUILD_CACHE_DIR
         echo 'Splitting test files into following parts'
-        split --numeric-suffixes -l `echo "($TOTAL_TESTS/4)+1" | bc` $TESTS_FILE_NAME $TESTS_FILES_PREFIX
+        split --numeric-suffixes -l `echo "($TOTAL_TESTS/5)+1" | bc` $TESTS_FILE_NAME $TESTS_FILES_PREFIX
         ls "$TESTS_FILES_PREFIX"* | sort | xargs wc -l
         sed -i 's/\\n/ /g' "$TESTS_FILES_PREFIX"*
         popd
-        ./gradlew compileTestScala --parallel
+        ./gradlew resolveAllDependencies --parallel
     - stage: 'build'
       name: 'Build Part 1'
       script: |
@@ -54,3 +54,7 @@ jobs:
         export TEST_FILES=`cat "$BUILD_CACHE_DIR/$TESTS_FILES_PREFIX"03`
         ./gradlew --no-daemon reportScoverage
       name: 'Build Part 4'
+    - script: |
+        export TEST_FILES=`cat "$BUILD_CACHE_DIR/$TESTS_FILES_PREFIX"04`
+        ./gradlew --no-daemon reportScoverage
+      name: 'Build Part 5'
